+++ presto-main/src/main/java/com/facebook/presto/operator/aggregation/AggregationImplementation.java
@@ -48,9 +48,9 @@ import java.util.stream.Stream;
 import static com.facebook.presto.operator.aggregation.AggregationMetadata.ParameterMetadata.ParameterType.BLOCK_INDEX;
 import static com.facebook.presto.operator.aggregation.AggregationMetadata.ParameterMetadata.ParameterType.STATE;
 import static com.facebook.presto.operator.aggregation.AggregationMetadata.ParameterMetadata.ParameterType.inputChannelParameterType;
-import static com.facebook.presto.operator.annotations.AnnotationHelpers.containsAnnotation;
-import static com.facebook.presto.operator.annotations.AnnotationHelpers.createTypeVariableConstraints;
-import static com.facebook.presto.operator.annotations.AnnotationHelpers.parseLiteralParameters;
+import static com.facebook.presto.operator.annotations.FunctionsParserHelper.containsAnnotation;
+import static com.facebook.presto.operator.annotations.FunctionsParserHelper.createTypeVariableConstraints;
+import static com.facebook.presto.operator.annotations.FunctionsParserHelper.parseLiteralParameters;
 import static com.facebook.presto.operator.annotations.ImplementationDependency.Factory.createDependency;
 import static com.facebook.presto.operator.annotations.ImplementationDependency.getImplementationDependencyAnnotation;
 import static com.facebook.presto.operator.annotations.ImplementationDependency.isImplementationDependencyAnnotation;
